+++ src/main/java/model/Pc.java
@@ -2,7 +2,7 @@ package model;
 
 public class Pc extends Character {
     private Inventory inventory;
-    public Pc(int health, int attack, String name,Inventory inventory,int goldAmount){
+    public Pc(int health, double attack, String name,Inventory inventory,int goldAmount){
         super(health,attack,name,goldAmount);
         this.inventory = inventory;
     }
